fixed merge issues add folders
