+++ resq/frontend/src/App.js
@@ -53,7 +53,7 @@ function App() {
             title: "Ongoing fire on Cami Sk.",
             desc: "A fire has been reported on Rumeli Hisarı Mh, Cami Sk, Sarıyer/İstanbul.",
             read: false
-        },{
+        }, {
             title: "New Soup Kitchen Near You",
             desc: "A soup kitchen has opened on Etiler, Hisar Üstü Nispetiye Cd No:1, Beşiktaş/İstanbul.",
             read: true
@@ -101,132 +101,113 @@ function App() {
         <QueryClientProvider client={queryClient}>
             <LocalizationProvider dateAdapter={AdapterDayjs}>
                 <ThemeProvider theme={theme}>
-                <Router>
-                    <div>
-                        <Navbar bg="light" variant="light" expand="lg">
-                            <Container ref={ref}>
-                                <Navbar.Brand href="/" style={{color: 'red', fontWeight: 'bold'}}>
-                                    <SmallRedCircle/>
-                                    ResQ
-                                </Navbar.Brand>
-                                <Navbar.Toggle aria-controls="basic-navbar-nav"/>
-                                <Navbar.Collapse id="basic-navbar-nav">
-                                    <Nav className="me-auto">
-                                        {navLinks.map(({path, label, icon}) => (
-                                            <Nav.Link key={path} href={path}>
-                                                {icon}
-                                                {label}
-                                            </Nav.Link>
-                                        ))}
-                                    </Nav>
-                                    <Nav className="ml-auto">
-                                        {token ?
-                                            <>
-                                                <Nav.Link key={"/account"} href={"/account"}
-                                                          style={{"marginLeft": "auto"}}>
-                                                    <SmallRedCircle/>
-                                                    <strong>Account</strong>
+                    <Router>
+                        <div>
+                            <Navbar bg="light" variant="light" expand="lg">
+                                <Container ref={ref}>
+                                    <Navbar.Brand href="/" style={{color: 'red', fontWeight: 'bold'}}>
+                                        <SmallRedCircle/>
+                                        ResQ
+                                    </Navbar.Brand>
+                                    <Navbar.Toggle aria-controls="basic-navbar-nav"/>
+                                    <Navbar.Collapse id="basic-navbar-nav">
+                                        <Nav className="me-auto">
+                                            {navLinks.map(({path, label, icon}) => (
+                                                <Nav.Link key={path} href={path}>
+                                                    {icon}
+                                                    {label}
                                                 </Nav.Link>
-                                                <Nav.Link key={"/requestcreation"} href={"/requestcreation"}
-                                                          style={{"marginLeft": "auto"}}>
-                                                    <SmallRedCircle/>
-                                                    <strong>Create Request</strong>
-                                                </Nav.Link>
-                                                <Nav.Link key={"/resourcecreation"} href={"/resourcecreation"}
-                                                          style={{"marginLeft": "auto"}}>
-                                                    <SmallRedCircle/>
-                                                    <strong>Create Resource</strong>
-                                                </Nav.Link>
-                                                <Nav.Link key={"/notifications"} href={"/notifications"}
-                                                          style={{"marginLeft": "auto"}}>
-                                                    <Badge badgeContent={notifications.filter(n => !n?.read).length}
-                                                           color={"primary"}>
-                                                        <NotificationsIcon/>
-                                                    </Badge>
-                                                </Nav.Link>
-                                                <Nav.Link key={"signout"} href={"#"} onClick={signOut}
-                                                          style={{"marginLeft": "auto"}}>
-                                                    <LogoutIcon/>
-                                                </Nav.Link>
-                                            </> :
-                                            <>
-                                                <Nav.Link key={'/signin'} href={'/signin'}
-                                                          style={{"marginLeft": "auto"}}>
-                                                    <strong>Sign In</strong>
-                                                </Nav.Link>
-                                                <Nav.Link key={'/signup'} href={'/signup'}
-                                                          style={{"marginLeft": "auto"}}>
-                                                    <strong>Sign Up</strong>
-                                                </Nav.Link>
+                                            ))}
+                                        </Nav>
+                                        <Nav className="ml-auto">
+                                            {token ?
+                                                <>
+                                                    <Nav.Link key={"/account"} href={"/account"}
+                                                              style={{"marginLeft": "auto"}}>
+                                                        <SmallRedCircle/>
+                                                        <strong>Account</strong>
+                                                    </Nav.Link>
+                                                    <Nav.Link key={"/requestcreation"} href={"/requestcreation"}
+                                                              style={{"marginLeft": "auto"}}>
+                                                        <SmallRedCircle/>
+                                                        <strong>Create Request</strong>
+                                                    </Nav.Link>
+                                                    <Nav.Link key={"/resourcecreation"} href={"/resourcecreation"}
+                                                              style={{"marginLeft": "auto"}}>
+                                                        <SmallRedCircle/>
+                                                        <strong>Create Resource</strong>
+                                                    </Nav.Link>
+                                                    <Nav.Link key={"/notifications"} href={"/notifications"}
+                                                              style={{"marginLeft": "auto"}}>
+                                                        <Badge badgeContent={notifications.filter(n => !n?.read).length}
+                                                               color={"primary"}>
+                                                            <NotificationsIcon/>
+                                                        </Badge>
+                                                    </Nav.Link>
+                                                    <Nav.Link key={"signout"} href={"#"} onClick={signOut}
+                                                              style={{"marginLeft": "auto"}}>
+                                                        <LogoutIcon/>
+                                                    </Nav.Link>
+                                                </> :
+                                                <>
+                                                    <Nav.Link key={'/signin'} href={'/signin'}
+                                                              style={{"marginLeft": "auto"}}>
+                                                        <strong>Sign In</strong>
+                                                    </Nav.Link>
+                                                    <Nav.Link key={'/signup'} href={'/signup'}
+                                                              style={{"marginLeft": "auto"}}>
+                                                        <strong>Sign Up</strong>
+                                                    </Nav.Link>
+                                                </>
+                                            }
+                                        </Nav>
+                                    </Navbar.Collapse>
+                                </Container>
+                            </Navbar>
+
+                            <main style={{height: `${height - 57}px`}}>
+                                <Routes>
+                                    {navLinks.map(({path, component}) => (
+                                        <Route key={path} path={path}
+                                               element={React.createElement(component, {
+                                                   token,
+                                                   setToken,
+                                                   role,
+                                                   setRole
+                                               })}/>
+                                    ))}
+                                    <Route path="/" element={<Navigate to="/map"/>}/>
+                                    <Route path="/rolerequest" state={{token, setToken}}
+                                           element={React.createElement(RoleRequest, {token, setToken})}/>
+                                    {
+                                        token ? <>
+                                                <Route path="/account" state={{token, setToken}}
+                                                       element={React.createElement(Account, {token, setToken})}/>
+                                                <Route path="/requestcreation" state={{token, setToken}}
+                                                       element={React.createElement(Request, {token, setToken})}/>
+                                                <Route path="/notifications" state={{token, setToken}}
+                                                       element={React.createElement(Notifications, {
+                                                           token,
+                                                           notifications,
+                                                           setNotifications
+                                                       })}/>
+                                                <Route path="/resourcecreation" state={{token, setToken}}
+                                                       element={React.createElement(Resource, {token, setToken})}/>
                                             </>
-                                        }
-                                    </Nav>
-                                </Navbar.Collapse>
-                            </Container>
-                        </Navbar>
-                        <main style={{height: `${height - 57}px`}}>
-                            <Routes>
-                                {navLinks.map(({path, component}) => (
-                                    <Route key={path} path={path}
-                                           element={React.createElement(component, {token, setToken, role, setRole})}/>
-                                ))}
-                                <Route path="/" element={<Navigate to="/map"/>}/>
-                                <Route path="/rolerequest" state={{token, setToken}}
-                                       element={React.createElement(RoleRequest, {token, setToken})}/>
-                                {
-                                    token ? <>
-                                            <Route path="/account" state={{token, setToken}}
-                                                   element={React.createElement(Account, {token, setToken})}/>
-                                            <Route path="/requestcreation" state={{token, setToken}}
-                                                   element={React.createElement(Request, {token, setToken})}/>
-                                            <Route path="/resourcecreation" state={{token, setToken}}
-                                                   element={React.createElement(Request, {token, setToken})}/>
-                                        </>
-                                        : <>
-                                            <Route path="/signin" state={{token, setToken}}
-                                                   element={React.createElement(SignIn, {token, setToken})}/>
-                                            <Route path="/signup" state={{token, setToken}}
-                                                   element={React.createElement(SignUp, {token, setToken})}/>
-                                        </>
+                                            : <>
+                                                <Route path="/signin" state={{token, setToken}}
+                                                       element={React.createElement(SignIn, {token, setToken})}/>
+                                                <Route path="/signup" state={{token, setToken}}
+                                                       element={React.createElement(SignUp, {token, setToken})}/>
+                                            </>
                                     }
-                                </Nav>
-                            </Navbar.Collapse>
-                        </Container>
-                    </Navbar>
-                    <main style={{ height: `${height - 57}px` }}>
-                        <Routes>
-                            {navLinks.map(({path, component}) => (
-                                <Route key={path} path={path}
-                                       element={React.createElement(component, {token, setToken, role, setRole})}/>
-                            ))}
-                            <Route path="/" element={<Navigate to="/map"/>}/>
-                            <Route path="/rolerequest" state={{token, setToken}}
-                                   element={React.createElement(RoleRequest, {token, setToken})}/>
-                            {
-                                token ? <>
-                                        <Route path="/account" state={{token, setToken}}
-                                               element={React.createElement(Account, {token, setToken})}/>
-                                        <Route path="/requestcreation" state={{token, setToken}}
-                                               element={React.createElement(Request, {token, setToken})}/>
-                                        <Route path="/notifications" state={{token, setToken}}
-                                               element={React.createElement(Notifications, {token, notifications, setNotifications})}/>
-                                    <Route path="/resourcecreation" state={{ token, setToken }}
-                                        element={React.createElement(Resource, { token, setToken })} />
-                                    </>
-                                    : <>
-                                        <Route path="/signin" state={{token, setToken}}
-                                               element={React.createElement(SignIn, {token, setToken})}/>
-                                        <Route path="/signup" state={{token, setToken}}
-                                               element={React.createElement(SignUp, {token, setToken})}/>
-                                    </>
-                            }
-                        </Routes>
-                    </main>
-                </div>
-            </Router>
-        </QueryClientProvider>
-        </ThemeProvider>
-    );
+                                </Routes>
+                            </main>
+                        </div>
+                    </Router>
+                </ThemeProvider>
+            </LocalizationProvider>
+        </QueryClientProvider>);
 }
 
 export default App;
