+++ resq/frontend/src/components/Request/RequestAddress.js
@@ -6,12 +6,10 @@ import Typography from '@mui/material/Typography';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import Checkbox from '@mui/material/Checkbox';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import disasterImage from '../disaster.png';
+import disasterImage from '../../disaster.png';
 import Avatar from '@mui/material/Avatar';
 import CssBaseline from '@mui/material/CssBaseline';
 import Container from '@mui/material/Container';
-import { ResourceContext } from './ResourceContext';
-import { useContext } from 'react';
 
 
 const customTheme = createTheme({
@@ -22,27 +20,19 @@ const customTheme = createTheme({
     },
 });
 
-export default function CreateResourceForm() {
-
-    const { resourceData, setResourceData } = useContext(ResourceContext);
-
-    const handleLongitudeChange = (longitudeValue) => {
-        setResourceData({
-            ...resourceData,
-            longitude: parseFloat(longitudeValue),
-        });
-    };
-
-    const handleLatitudeChange = (latitudeValue) => {
-        setResourceData({
-            ...resourceData,
-            latitude: parseFloat(latitudeValue),
-        });
-    };
+export default function CreateRequestForm() {
+    const [request, setRequest] = useState({
+        type: '',
+        status: '',
+        urgency: '',
+        location: '',
+        quantity: '',
+        description: '',
+    });
 
     const handleChange = (event) => {
         const { name, value } = event.target;
-        setResource(prevState => ({
+        setRequest(prevState => ({
             ...prevState,
             [name]: value
         }));
@@ -50,7 +40,8 @@ export default function CreateResourceForm() {
 
     const handleSubmit = (event) => {
         event.preventDefault();
-        console.log(resource);
+        console.log(request);
+        //send this data to a backend server
     };
 
     return (
@@ -67,10 +58,32 @@ export default function CreateResourceForm() {
                 >
 
                     <Typography component="h1" variant="h5" sx={{ color: 'red', fontWeight: 'bold', margin: '0' }}>
-                        Resource Delivery Address
+                        Create Request
                     </Typography>
                     <React.Fragment>
                         <Grid container spacing={3}>
+                            <Grid item xs={12} sm={6}>
+                                <TextField
+                                    required
+                                    id="firstName"
+                                    name="firstName"
+                                    label="First name"
+                                    fullWidth
+                                    autoComplete="given-name"
+                                    variant="standard"
+                                />
+                            </Grid>
+                            <Grid item xs={12} sm={6}>
+                                <TextField
+                                    required
+                                    id="lastName"
+                                    name="lastName"
+                                    label="Last name"
+                                    fullWidth
+                                    autoComplete="family-name"
+                                    variant="standard"
+                                />
+                            </Grid>
                             <Grid item xs={12}>
                                 <TextField
                                     required
@@ -78,7 +91,7 @@ export default function CreateResourceForm() {
                                     name="address1"
                                     label="Address line 1"
                                     fullWidth
-                                    autoComplete="shipping address-line1"
+                                    autoComplete="address1"
                                     variant="standard"
                                 />
                             </Grid>
@@ -88,7 +101,7 @@ export default function CreateResourceForm() {
                                     name="address2"
                                     label="Address line 2"
                                     fullWidth
-                                    autoComplete="shipping address-line2"
+                                    autoComplete="address2"
                                     variant="standard"
                                 />
                             </Grid>
@@ -99,7 +112,7 @@ export default function CreateResourceForm() {
                                     name="city"
                                     label="City"
                                     fullWidth
-                                    autoComplete="shipping address-level2"
+                                    autoComplete="city"
                                     variant="standard"
                                 />
                             </Grid>
@@ -115,6 +128,17 @@ export default function CreateResourceForm() {
                             <Grid item xs={12} sm={6}>
                                 <TextField
                                     required
+                                    id="zip"
+                                    name="zip"
+                                    label="Zip / Postal code"
+                                    fullWidth
+                                    autoComplete="shipping postal-code"
+                                    variant="standard"
+                                />
+                            </Grid>
+                            <Grid item xs={12} sm={6}>
+                                <TextField
+                                    required
                                     id="country"
                                     name="country"
                                     label="Country"
@@ -123,21 +147,10 @@ export default function CreateResourceForm() {
                                     variant="standard"
                                 />
                             </Grid>
-                            <Grid item xs={12} sm={6}>
-                                <TextField
-                                    required
-                                    id="nameofplace"
-                                    name="nameofplace"
-                                    label="Name of the Delivery Place"
-                                    fullWidth
-                                    autoComplete="delivery place"
-                                    variant="standard"
-                                />
-                            </Grid>
                             <Grid item xs={12}>
                                 <FormControlLabel
                                     control={<Checkbox color="error" name="saveAddress" value="yes" />}
-                                    label="I have delivered my resources to the above mentioned address."
+                                    label="Use this address for my request"
                                 />
                             </Grid>
                         </Grid>
