+++ resq/frontend/src/components/Resource/ResourceDetail1.js
