+++ resq/frontend/src/components/Resource/ResourceDetail2.js
